requirements.txt
cryptography
Flask
Flask-SocketIO
pycryptodome

